feat:refactor timepicker
