+++ src/components/date-picker/DatePicker.tsx
@@ -56,14 +56,9 @@ function DatePicker({
             : new Date()
     )
 
-    console.log("new\n", newMinDate, "\n", newMaxDate, new Date());
-
-
-    console.log("newDisable\n", newDisablePastDates, newMinDate, newMinDate.getFullYear());
     const startYear = newMinDate.getFullYear()
     const endYear = newMaxDate.getFullYear()
     const yearsPerPage = 12
-    console.log("Year\n", startYear, endYear, yearsPerPage, newMaxDate, newMaxDate.getFullYear());
 
     const today = useMemo(() => {
         const d = new Date()
