+++ src/components/modal/BookingSelect.tsx
@@ -5,7 +5,7 @@ import AlertConfirm from "./AlertConfirm"
 import PrimaryButton from "../buttons/PrimaryButton"
 import { useDatePicker } from "@/hooks/useDatePicker"
 import { useTimePicker } from "@/hooks/useTimePicker"
-import TimePicker from "../time-picker/TimePicker"
+import TimePicker from "../time-picker/TimePickerOld"
 import DatePicker from "../date-picker/DatePicker"
 import { Calendar, Clock } from 'lucide-react';
